feat: add rest and default
