--- src/[[...slug]]/client.tsx
@@ -1,13 +1,0 @@
-// 'use client';
-
-// import React from 'react';
-
-// import dynamic from 'next/dynamic';
-
-// import StoreProvider from '@/app/StoreProvider';
-
-// const App = dynamic(() => import('@/views/HomeView/HomeView'), { ssr: false });
-
-// export function ClientOnly() {
-//   return <App />;
-// }
