--- src/[[...slug]]/page.tsx
@@ -1,12 +1,0 @@
-// import { ClientOnly } from '@/app/[[...slug]]/client';
-
-// import '@/styles/index.css';
-
-// // eslint-disable-next-line react-refresh/only-export-components
-// export function generateStaticParams() {
-//   return [{ slug: [''] }, { slug: ['details', '1'] }];
-// }
-
-// export default function Page() {
-//   return <ClientOnly />;
-// }
