+++ src/context/ThemeContext.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { createContext } from 'react';
 
 interface TypeThemeContext {
